create generate function
