+++ datamodel_code_generator/__main__.py
@@ -4,32 +4,25 @@
 Main function.
 """
 
-import contextlib
 import json
-import os
 import signal
 import sys
 from argparse import ArgumentParser, FileType, Namespace
 from collections import defaultdict
-from datetime import datetime, timezone
-from enum import Enum, IntEnum
+from enum import IntEnum
 from pathlib import Path
-from typing import IO, Any, DefaultDict, Dict, Iterator, Optional, Sequence, Type
+from typing import Any, DefaultDict, Dict, Optional, Sequence
 
 import argcomplete
 
 from datamodel_code_generator import (
+    DEFAULT_BASE_CLASS,
+    Error,
+    InputFileType,
     PythonVersion,
     enable_debug_message,
-    load_json_or_yaml,
+    generate,
 )
-from datamodel_code_generator.model.pydantic import (
-    BaseModel,
-    CustomRootType,
-    dump_resolve_reference_action,
-)
-from datamodel_code_generator.model.pydantic.base_model import DataModelField
-from datamodel_code_generator.parser.base import Parser
 
 
 class Exit(IntEnum):
@@ -40,38 +33,11 @@ class Exit(IntEnum):
     KeyboardInterrupt = 2
 
 
-class InputFileType(Enum):
-    Auto = 'auto'
-    OpenAPI = 'openapi'
-    JsonSchema = 'jsonschema'
-    Json = 'json'
-    Yaml = 'yaml'
-
-
 def sig_int_handler(_: int, __: Any) -> None:  # pragma: no cover
     exit(Exit.OK)
 
 
 signal.signal(signal.SIGINT, sig_int_handler)
-
-
-@contextlib.contextmanager
-def chdir(path: Optional[Path]) -> Iterator[None]:
-    """Changes working directory and returns to previous on exit."""
-
-    if path is None:
-        yield
-    else:
-        prev_cwd = Path.cwd()
-        try:
-            os.chdir(path if path.is_dir() else path.parent)
-            yield
-        finally:
-            os.chdir(prev_cwd)
-
-
-def is_openapi(text: str) -> bool:
-    return 'openapi' in load_json_or_yaml(text)
 
 
 arg_parser = ArgumentParser()
@@ -92,7 +58,7 @@ arg_parser.add_argument(
     '--base-class',
     help='Base Class (default: pydantic.BaseModel)',
     type=str,
-    default='pydantic.BaseModel',
+    default=DEFAULT_BASE_CLASS,
 )
 arg_parser.add_argument(
     '--custom-template-dir', help='Custom Template Directory', type=str
@@ -142,100 +108,24 @@ def main(args: Optional[Sequence[str]] = None) -> Exit:
     else:
         extra_template_data = None
 
-    text: str = namespace.input.read()
-
-    input_file_type = InputFileType(namespace.input_file_type)
-
-    if input_file_type == InputFileType.Auto:
-        try:
-            input_file_type = (
-                InputFileType.OpenAPI if is_openapi(text) else InputFileType.JsonSchema
-            )
-        except:
-            print('Invalid file format')
-            return Exit.ERROR
-
-    if input_file_type == InputFileType.OpenAPI:
-        from datamodel_code_generator.parser.openapi import OpenAPIParser
-
-        parser_class: Type[Parser] = OpenAPIParser
-    else:
-        from datamodel_code_generator.parser.jsonschema import JsonSchemaParser
-
-        parser_class = JsonSchemaParser
-
-        if input_file_type in [InputFileType.Json, InputFileType.Yaml]:
-            try:
-                if input_file_type == InputFileType.Json:
-                    obj: Dict = json.loads(text)
-                else:
-                    import yaml
-
-                    obj = yaml.safe_load(text)
-            except:
-                print('Invalid file format')
-                return Exit.ERROR
-            from genson import SchemaBuilder
-
-            builder = SchemaBuilder()
-            builder.add_object(obj)
-            text = json.dumps(builder.to_schema())
-
-    parser = parser_class(
-        BaseModel,
-        CustomRootType,
-        DataModelField,
-        base_class=namespace.base_class,
-        custom_template_dir=namespace.custom_template_dir,
-        extra_template_data=extra_template_data,
-        target_python_version=PythonVersion(namespace.target_python_version),
-        text=text,
-        dump_resolve_reference_action=dump_resolve_reference_action,
-        validation=namespace.validation,
-    )
-
-    output = Path(namespace.output) if namespace.output is not None else None
-
-    with chdir(output):
-        result = parser.parse()
-
-    if isinstance(result, str):
-        modules = {output: result}
-    else:
-        if output is None:
-            print('Modular references require an output directory')
-            return Exit.ERROR
-        if output.suffix:
-            print('Modular references require an output directory, not a file')
-            return Exit.ERROR
-        modules = {
-            output.joinpath(*name): body for name, body in sorted(result.items())
-        }
-
-    timestamp = datetime.now(timezone.utc).replace(microsecond=0).isoformat()
-    header = f'''\
-# generated by datamodel-codegen:
-#   filename:  {Path(namespace.input.name).name}
-#   timestamp: {timestamp}'''
-
-    file: Optional[IO[Any]]
-    for path, body in modules.items():
-        if path is not None:
-            if not path.parent.exists():
-                path.parent.mkdir(parents=True)
-            file = path.open('wt')
-        else:
-            file = None
-
-        print(header, file=file)
-        if body:
-            print('', file=file)
-            print(body.rstrip(), file=file)
-
-        if file is not None:
-            file.close()
-
-    return Exit.OK
+    try:
+        generate(
+            input_name=namespace.input.name,
+            input_text=namespace.input.read(),
+            input_file_type=InputFileType(namespace.input_file_type),
+            output=Path(namespace.output) if namespace.output is not None else None,
+            target_python_version=PythonVersion(namespace.target_python_version),
+            base_class=namespace.base_class,
+            custom_template_dir=namespace.custom_template_dir,
+            extra_template_data=extra_template_data,
+            validation=namespace.validation
+        )
+        return Exit.OK
+    except Error as e:
+        print(str(e), file=sys.stderr)
+        return Exit.ERROR
+    except:
+        return Exit.ERROR
 
 
 if __name__ == '__main__':
